refactor(core): split core service into layered services
- add Ctx as the shared context and ChainService as the single chain seam
- extract the startup contract handshake into ChainDeployment
- split the CoreService god-object into domain services behind accessors
- move cross-service primitives into service/shared with *Ops naming
- fold cycle sweeps into ClearingService
- make CoreService a composition root with no domain logic
- update tests to call through the service accessors
