+++ core/src/ethereum/mod.rs
@@ -1,11 +1,15 @@
+pub mod chain;
 pub mod contract;
+pub mod deployment;
 pub mod event_data;
 pub mod event_handler;
 pub mod proxy;
 pub mod revert;
 pub mod scanner;
 
+pub use chain::ChainService;
 pub use contract::contract_abi;
+pub use deployment::ChainDeployment;
 pub use proxy::{
     ClearingCommitInput, ClearingCycleView, ClearingTxResult, CoreContractApi, CoreContractProxy,
     CreditorSettlement, DebtorSettlement, GuaranteeVersionConfig,
